pyproject.toml
[tool.poetry]
name = "car_search"
version = "0.1.0"
description = "Intelligent Car Shopping with LLM"
authors = ["Your Name <you@example.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.8"
beautifulsoup4 = "4.12.2"
requests = "2.31.0"
selenium = "4.19.0"
webdriver-manager = "4.0.1"
aiohttp = "3.9.3"
pydantic = "2.5.3"
textual = "0.43.1"
rich = "13.7.0"
google-generativeai = "0.3.2"
openai = "1.76.0"
anthropic = "0.8.1"
python-dotenv = "1.0.0"
typer = "0.9.0"

[tool.poetry.dev-dependencies]
pytest = "7.4.3"
responses = "0.24.1"
pytest-mock = "3.12.0"
pytest-cov = "4.1.0"

[build-system]
requires = ["poetry-core>=1.2.0"]
build-backend = "poetry.core.masonry.api" 